+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -75,7 +75,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-02-25T12:28:13.397Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-02-26T10:55:48.615Z")
     r"""The date and time the resource was created"""